no_requirements.txt
absl-py==1.3.0
adjustText==0.8
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.1
anndata==0.8.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1667325728359/work
async-timeout==4.0.2
attrs==22.1.0
autopage==0.5.1
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
bamnostic==1.1.10
beautifulsoup4==4.12.2
bertviz==1.4.0
bioinfokit==2.1.4
biomart==0.9.2
biopython==1.83
bleach==6.0.0
boto3==1.26.104
botocore==1.29.104
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chex==0.1.5
click==8.1.7
cliff==4.1.0
cmaes==0.9.0
cmd2==2.4.2
colorlog==6.7.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1668713578881/work
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
datasets==2.11.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1666826398899/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
decoupler==1.5.0
deepspeed==0.8.0
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.6
dm-tree==0.1.7
docrep==0.3.2
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
episcanpy @ git+https://github.com/colomemaria/epiScanpy@0498043252c58b4c4187007c215dd6d2143d121b
et-xmlfile==1.1.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fastjsonschema==2.16.3
filelock==3.10.7
flax==0.6.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2022.11.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
greenlet==2.0.1
grpcio==1.50.0
gseapy==1.0.5
h5py==3.7.0
harmonypy==0.0.9
hdbscan==0.8.29
hjson==3.1.0
HTSeq==2.0.3
huggingface-hub==0.13.3
idna==3.4
igraph==0.10.2
importlib-metadata==4.13.0
inflect==7.0.0
intervaltree==3.1.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1669056400400/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1667140637743/work
ipython-genutils==0.2.0
ipywidgets==8.0.2
isoduration==20.11.0
jax==0.3.25
jaxlib==0.3.25
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1668623095912/work
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kneed==0.8.5
legacy-api-wrap==1.4
leidenalg==0.9.0
llvmlite==0.39.1
loompy==3.0.7
louvain==0.8.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
matplotlib-venn==0.11.10
mistune==2.0.5
ml-collections==0.1.1
mlxtend==0.23.0
msgpack==1.0.4
mudata==0.2.1
MulticoreTSNE==0.1
multidict==6.0.2
multipledispatch==0.6.0
multiprocess==0.70.14
natsort==8.2.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==2.8.8
ninja==1.11.1
notebook==6.5.4
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
numpy-groupies==0.11.2
numpyro==0.10.1
oauthlib==3.2.2
omnipath==1.0.7
openpyxl==3.0.10
opt-einsum==3.3.0
optax==0.1.4
optuna==3.0.5
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.5.2
pandocfilters==1.5.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
patsy==0.5.3
pbr==5.11.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1657729053205/work
plotly==5.13.1
prettytable==3.5.0
prometheus-client==0.16.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1669057097528/work
protobuf==3.20.3
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1667885909637/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybedtools==0.9.0
pycparser==2.21
pydantic==1.10.4
pyDeprecate==0.3.2
pydeseq2==0.3.5
pydot==1.4.2
pyg-lib==0.1.0+pt112cu113
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1660666458521/work
pyjaspar==3.0.0
pynndescent==0.5.8
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
pyperclip==1.8.2
pyro-api==0.1.2
pyro-ppl==1.8.3
pyrsistent==0.19.3
pysam==0.21.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-json-logger==2.0.7
pytorch-lightning==1.7.7
pytorch-pretrained-bert==0.6.2
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.2
QtPy==2.3.1
regex==2023.3.23
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.9
s3transfer==0.6.0
sanbomics==0.1.0
scanpy==1.9.1
scib==1.1.4
scikit-learn==1.1.3
scikit-misc==0.3.0
scipy==1.11.2
scvelo==0.3.2
scvi-tools==0.19.0
seaborn==0.12.1
Send2Trash==1.8.0
sentencepiece==0.1.97
session-info==1.0.0
setuptools-scm==8.0.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.45
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1668260892666/work
statsmodels==0.13.5
stdlib-list==0.8.0
stevedore==4.1.1
strandex==0.4.1
SUPPA==2.3
tabulate==0.9.0
tbb==2021.11.0
tenacity==8.2.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorstore==0.1.28
terminado==0.17.1
texttable==1.6.7
textwrap3==0.9.2
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1+cu113
torch-geometric==2.2.0
torch-scatter==2.1.0+pt112cu113
torch-sparse==0.6.16+pt112cu113
torchaudio==0.12.1+cu113
torchmetrics==0.10.3
torchvision==0.13.1+cu113
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1666788587690/work
tqdm==4.64.1
traitlets==5.9.0
transformers==4.27.4
typing_extensions==4.4.0
umap-learn==0.5.3
UpSetPlot==0.9.0
uri-template==1.2.0
urllib3==1.26.12
velocyto==0.17.17
venn==0.1.3
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.15.0
xxhash==3.2.0
yarl==1.8.1
zipp==3.10.0
